fix circular dependency error
