+++ clustpy/deep/_train_utils.py
@@ -1,10 +1,6 @@
 from clustpy.deep.autoencoders import FeedforwardAutoencoder
-from clustpy.deep import get_dataloader, detect_device, encode_batchwise
 import torch
 import copy
-import numpy as np
-from sklearn.base import ClusterMixin
-import inspect
 
 
 def _get_default_layers(input_dim: int, embedding_size: int) -> list:
@@ -87,142 +83,3 @@ def get_trained_autoencoder(trainloader: torch.utils.data.DataLoader, optimizer_
         autoencoder = copy.deepcopy(autoencoder)
 
     return autoencoder
-
-
-def run_initial_clustering(X: np.ndarray, n_clusters: int, clustering_class: ClusterMixin, clustering_params: dict,
-                           random_state: np.random.RandomState) -> (int, np.ndarray, np.ndarray, ClusterMixin):
-    """
-    Get an initial clustering result for a deep clustering algorithm.
-    This result can then be refined by the optimization of the autoencoder.
-
-    Parameters
-    ----------
-    X : np.ndarray
-        the given data set
-    n_clusters : int
-        number of clusters. Can be None if a corresponding initial_clustering_class is given, e.g. DBSCAN
-    clustering_class : ClusterMixin
-        the class of the initial clustering algorithm
-    clustering_params : dict
-        the parameters for the initial clustering algorithm
-    random_state : np.random.RandomState
-        use a fixed random state to get a repeatable solution
-
-    Returns
-    -------
-    tuple : (int, np.ndarray, np.ndarray, ClusterMixin)
-        The number of clusters (can change if e.g. DBSCAN is used),
-        The initial cluster labels,
-        The initial cluster centers,
-        The clustering object
-    """
-    # Get possible input parameters of the clustering algorithm
-    clustering_class_parameters = inspect.getfullargspec(clustering_class).args + inspect.getfullargspec(
-        clustering_class).kwonlyargs
-    # Check if n_clusters or n_components is contained in the possible parameters
-    if "n_clusters" in clustering_class_parameters:
-        if "random_state" in clustering_class_parameters:
-            clustering_algo = clustering_class(n_clusters=n_clusters, random_state=random_state, **clustering_params)
-        else:
-            clustering_algo = clustering_class(n_clusters=n_clusters, **clustering_params)
-    elif "n_components" in clustering_class_parameters:  # in case of GMM
-        if "random_state" in clustering_class_parameters:
-            clustering_algo = clustering_class(n_components=n_clusters, random_state=random_state, **clustering_params)
-        else:
-            clustering_algo = clustering_class(n_components=n_clusters, **clustering_params)
-    else:  # in case of e.g., DBSCAN
-        if "random_state" in clustering_class_parameters:
-            clustering_algo = clustering_class(random_state=random_state, **clustering_params)
-        else:
-            clustering_algo = clustering_class(**clustering_params)
-    # Run algorithm
-    clustering_algo.fit(X)
-    # Check if clustering algorithm return cluster centers
-    if hasattr(clustering_algo, "cluster_centers_"):
-        labels = clustering_algo.labels_
-        centers = clustering_algo.cluster_centers_
-    elif hasattr(clustering_algo, "means_"):  # in case of GMM
-        labels = clustering_algo.predict(X)
-        centers = clustering_algo.means_
-    else:  # in case of e.g., DBSCAN
-        labels = clustering_algo.labels_
-        centers = np.array([np.mean(X[labels == i], axis=0) for i in np.unique(labels) if i >= 0])
-    n_clusters = np.sum(np.unique(labels) >= 0)  # Needed for DBSCAN, XMeans, GMeans, ...
-    return n_clusters, labels, centers, clustering_algo
-
-
-def get_standard_initial_deep_clustering_setting(X: np.ndarray, n_clusters: int, batch_size: int,
-                                                 pretrain_optimizer_params: dict, pretrain_epochs: int,
-                                                 optimizer_class: torch.optim.Optimizer,
-                                                 loss_fn: torch.nn.modules.loss._Loss,
-                                                 autoencoder: torch.nn.Module, embedding_size: int,
-                                                 custom_dataloaders: tuple, initial_clustering_class: ClusterMixin,
-                                                 initial_clustering_params: dict,
-                                                 random_state: np.random.RandomState,
-                                                 autoencoder_class: torch.nn.Module = FeedforwardAutoencoder) -> (
-        torch.device, torch.utils.data.DataLoader, torch.utils.data.DataLoader, torch.nn.Module, np.ndarray, int,
-        np.ndarray, np.ndarray, ClusterMixin):
-    """
-    Get the initial setting for most deep clustering algorithms by pretraining an autoencoder and obtraining an initial clustering result.
-    This function further returns the device, where the optimization should take place (e.g., CPU or GPU), and the dataloaders.
-
-    Parameters
-    ----------
-    X : np.ndarray / torch.Tensor
-        the given data set. Can be a np.ndarray or a torch.Tensor
-    n_clusters : int
-        number of clusters. Can be None if a corresponding initial_clustering_class is given, e.g. DBSCAN
-    batch_size : int
-        size of the data batches
-    pretrain_optimizer_params : dict
-        parameters of the optimizer for the pretraining of the autoencoder, includes the learning rate
-    pretrain_epochs : int
-        number of epochs for the pretraining of the autoencoder
-    optimizer_class : torch.optim.Optimizer
-        the optimizer
-    loss_fn : torch.nn.modules.loss._Loss
-         loss function for the reconstruction
-    autoencoder : torch.nn.Module
-        the input autoencoder. If None a new FeedforwardAutoencoder will be created
-    embedding_size : int
-        size of the embedding within the autoencoder
-    custom_dataloaders : tuple
-        tuple consisting of a trainloader (random order) at the first and a test loader (non-random order) at the second position.
-        If None, the default dataloaders will be used
-    initial_clustering_class : ClusterMixin
-        clustering class to obtain the initial cluster labels after the pretraining
-    initial_clustering_params : dict
-        parameters for the initial clustering class
-    random_state : np.random.RandomState
-        use a fixed random state to get a repeatable solution
-    autoencoder_class : torch.nn.Module
-        The autoencoder class that should be used (default: FeedforwardAutoencoder)
-
-    Returns
-    -------
-    tuple : (torch.device, torch.utils.data.DataLoader, torch.utils.data.DataLoader, torch.nn.Module, np.ndarray, int, np.ndarray, np.ndarray, ClusterMixin)
-        The device,
-        The trainloader,
-        The testloader,
-        The pretrained autoencoder,
-        The embedded data,
-        The number of clusters (can change if e.g. DBSCAN is used),
-        The initial cluster labels,
-        The initial cluster centers,
-        The clustering object
-    """
-    device = detect_device()
-    if custom_dataloaders is None:
-        trainloader = get_dataloader(X, batch_size, True, False)
-        testloader = get_dataloader(X, batch_size, False, False)
-    else:
-        trainloader, testloader = custom_dataloaders
-    autoencoder = get_trained_autoencoder(trainloader, pretrain_optimizer_params, pretrain_epochs, device,
-                                          optimizer_class, loss_fn, embedding_size, autoencoder, autoencoder_class)
-    # Execute initial clustering in embedded space
-    embedded_data = encode_batchwise(testloader, autoencoder, device)
-    n_clusters, init_labels, init_centers, init_cluster_obj = run_initial_clustering(embedded_data, n_clusters,
-                                                                                     initial_clustering_class,
-                                                                                     initial_clustering_params,
-                                                                                     random_state)
-    return device, trainloader, testloader, autoencoder, embedded_data, n_clusters, init_labels, init_centers, init_cluster_obj
