+++ clustpy/deep/dkm.py
@@ -6,7 +6,7 @@ Collin Leiber
 from clustpy.deep._utils import encode_batchwise, squared_euclidean_distance, predict_batchwise, \
     set_torch_seed, embedded_kmeans_prediction
 from clustpy.deep._data_utils import get_dataloader, augmentation_invariance_check
-from clustpy.deep._train_utils import get_standard_initial_deep_clustering_setting
+from clustpy.deep.dec import get_standard_initial_deep_clustering_setting
 import torch
 import numpy as np
 from sklearn.cluster import KMeans
